+++ trans/src/sparc/local/szs_als.h
@@ -15,59 +15,59 @@
     BASIC SIZES AND ALIGNMENTS
 */
 
-#define TOP_SZ			( ( long ) 0 )
+#define TOP_SZ			0
 #define TOP_ALIGN		const_al1
 
-#define PROC_SZ			( ( long ) 32 )
+#define PROC_SZ			32
 #define PROC_ALIGN		const_al32
 
-#define LLVALUE_SZ		( ( long ) 32 )
+#define LLVALUE_SZ		32
 #define LLVALUE_ALIGN		const_al32
 
-#define PTR_SZ			( ( long ) 32 )
+#define PTR_SZ			32
 #define PTR_ALIGN		const_al32
-#define PTRBIT_SZ		( ( long ) 32 )
+#define PTRBIT_SZ		32
 
-#define OFFSET_SZ		( ( long ) 32 )
+#define OFFSET_SZ		32
 #define OFFSET_ALIGN		const_al32
 
 #define BF_ALIGN		const_al1
 #define STRUCT_ALIGN		const_al8
 
-#define SHREAL_SZ		( ( long ) 32 )
+#define SHREAL_SZ		32
 #define SHREAL_ALIGN		const_al32
 
-#define REAL_SZ			( ( long ) 64 )
+#define REAL_SZ			64
 #define REAL_ALIGN		const_al64
 
-#define DOUBLE_SZ		( (has & HAS_LONG_DOUBLE) ? ( long ) 128 : ( long ) 64 )
+#define DOUBLE_SZ		( (has & HAS_LONG_DOUBLE) ? 128 : 64 )
 #define DOUBLE_ALIGN		const_al64
 
-#define UCHAR_SZ		( ( long ) 8 )
+#define UCHAR_SZ		8
 #define UCHAR_ALIGN		const_al8
 
-#define SCHAR_SZ		( ( long ) 8 )
+#define SCHAR_SZ		8
 #define SCHAR_ALIGN		const_al8
 
-#define UWORD_SZ		( ( long ) 16 )
+#define UWORD_SZ		16
 #define UWORD_ALIGN		const_al16
 
-#define SWORD_SZ		( ( long ) 16 )
+#define SWORD_SZ		16
 #define SWORD_ALIGN		const_al16
 
-#define ULONG_SZ		( ( long ) 32 )
+#define ULONG_SZ		32
 #define ULONG_ALIGN		const_al32
 
-#define U64_SZ			( ( long ) 64 )
+#define U64_SZ			64
 #define U64_ALIGN		const_al32
 
-#define S64_SZ			( ( long ) 64 )
+#define S64_SZ			64
 #define S64_ALIGN		const_al32
 
-#define SLONG_SZ		( ( long ) 32 )
+#define SLONG_SZ		32
 #define SLONG_ALIGN		const_al32
 
-#define BF_STORE_UNIT		( ( long ) 32 )
+#define BF_STORE_UNIT		32
 #define MAX_BF_SIZE		const_al32
 
 #define ALLOCA_ALIGN		const_al64
